chore: Move EngineStatus definition to inference_server_management.proto
This is to avoid cyclic dependency in upcoming change.
