+++ ts/api/v1/inference_server_management.pb.ts
@@ -5,7 +5,26 @@
 */
 
 import * as fm from "../../fetch.pb"
-import * as LlmarinerInferenceServerV1Inference_server_worker from "./inference_server_worker.pb"
+export type EngineStatusSyncStatus = {
+  in_progress_model_ids?: string[]
+}
+
+export type EngineStatusModel = {
+  id?: string
+  is_ready?: boolean
+  in_progress_task_count?: number
+  gpu_allocated?: number
+}
+
+export type EngineStatus = {
+  engine_id?: string
+  model_ids?: string[]
+  sync_status?: EngineStatusSyncStatus
+  ready?: boolean
+  models?: EngineStatusModel[]
+  cluster_id?: string
+}
+
 export type InferenceStatus = {
   cluster_statuses?: ClusterStatus[]
   task_status?: TaskStatus
@@ -18,7 +37,7 @@ export type TaskStatus = {
 export type ClusterStatus = {
   id?: string
   name?: string
-  engine_statuses?: LlmarinerInferenceServerV1Inference_server_worker.EngineStatus[]
+  engine_statuses?: EngineStatus[]
   model_count?: number
   in_progress_task_count?: number
   gpu_allocated?: number
